pack/requirements.txt
alabaster==0.7.9
anaconda-clean==1.0
anaconda-client==1.5.1
anaconda-navigator==1.3.1
appdirs==1.4.0
argcomplete==1.0.0
astroid==1.4.7
astropy==1.2.1
attrs==17.3.0
Automat==0.6.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.1
bitarray==0.8.1
blaze==0.10.1
bokeh==0.12.2
boto==2.42.0
Bottleneck==1.1.0
certifi==2017.7.27.1
cffi==1.7.0
chardet==3.0.4
chest==0.2.3
click==6.6
cloudpickle==0.2.1
clyent==1.2.2
colorama==0.3.9
coloredlogs==8.0
comtypes==1.1.2
conda==4.2.9
conda-build==2.0.2
configobj==5.0.6
constantly==15.1.0
contextlib2==0.5.3
cryptography==1.5
cssselect==1.0.1
cycler==0.10.0
Cython==0.24.1
cytoolz==0.8.0
dask==0.11.0
datashape==0.5.2
decorator==4.0.10
defusedxml==0.5.0
dill==0.2.5
dj-database-url==0.4.2
Django==1.11.5
django-allauth==0.33.0
djangorestframework==3.6.4
docutils==0.12
dynd===c328ab7
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.6
Flask==0.11.1
Flask-Cors==2.1.2
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.7.1
h5py==2.6.0
HeapDict==1.0.0
httpie==0.9.9
httpie-django-auth==0.1.1
humanfriendly==4.7
hyperlink==17.3.1
idna==2.6
imagesize==0.7.1
incremental==17.5.0
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
Keras==1.1.1
lazy-object-proxy==1.2.1
llvmlite==0.13.0
locket==0.2.0
lxml==3.6.4
MarkupSafe==0.23
matplotlib==1.5.3
menuinst==1.4.1
mistune==0.7.3
mpmath==0.19
multipledispatch==0.4.8
nb-anacondacloud==1.2.0
nb-conda==2.0.0
nb-conda-kernels==2.0.0
nbconvert==4.2.0
nbformat==4.1.0
nbpresent==3.0.2
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.2.3
numba==0.28.1
numexpr==2.6.1
numpy==1.12.0
oauthlib==2.0.3
odo==0.5.0
opencv-python==3.2.0.7
openpyxl==2.3.2
packaging==16.8
pandas==0.18.1
parsel==1.2.0
partd==0.3.6
path.py==0.0.0
pathlib2==2.1.0
patsy==0.4.1
pep8==1.7.0
pickleshare==0.7.4
Pillow==3.3.1
pkginfo==1.3.2
ply==3.9
prompt-toolkit==1.0.3
protobuf==3.1.0
psutil==4.3.1
psycopg2==2.6.2
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.2.1
pyautogui==0.9.33
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
PyDispatcher==2.0.5
pyflakes==1.3.0
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.5.4
pymsgbox==1.0.4
pynput==1.3.9
pyOpenSSL==16.0.0
pyparsing==2.1.10
pyreadline==2.1
pyscreeze==0.1.9
pytest==2.9.2
python-dateutil==2.5.3
python3-openid==3.1.0
pytweening==1.0.3
pytz==2017.2
pywin32==220
PyYAML==3.12
pyzmq==15.4.0
QtAwesome==0.3.3
qtconsole==4.2.1
QtPy==1.1.2
queuelib==1.4.2
requests==2.18.4
requests-oauthlib==0.8.0
rope-py3k==0.9.4.post1
ruamel-yaml===-VERSION
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.18.1
Scrapy==1.4.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sockjs-tornado==1.0.3
sphinx==1.4.6
spyder==3.0.0
SQLAlchemy==1.0.13
statsmodels==0.6.1
sympy==1.0
tables==3.2.2
tensorflow==0.11.0
Theano==0.8.2
toolz==0.8.0
tornado==4.4.1
traitlets==4.3.0
Twisted==17.9.0
unicodecsv==0.14.1
urllib3==1.22
virtualenv==15.1.0
w3lib==1.18.0
wcwidth==0.1.7
Werkzeug==0.11.11
whitenoise==3.3.1
widgetsnbextension==1.2.6
win-unicode-console==0.5
wrapt==1.10.6
xlrd==1.0.0
XlsxWriter==0.9.3
xlwings==0.10.0
xlwt==1.1.2
zope.interface==4.4.3
